Making praat and parselmouth compile on MSVC

+++ praat/sys/Graphics_record.cpp
@@ -175,7 +175,7 @@ void Graphics_play (Graphics me, Graphics thee) {
 								0, nrow - 1, y1, y2, minimum, maximum);
 				Melder_free (z);
 			}  break;
-			case SET_FONT: Graphics_setFont (thee, (enum kGraphics_font) get); break;
+			case SET_FONT: Graphics_setFont (thee, (enum kGraphics_font) (int) get); break;
 			case SET_FONT_SIZE: Graphics_setFontSize (thee, (int) get); break;
 			case SET_FONT_STYLE: Graphics_setFontStyle (thee, (int) get); break;
 			case SET_TEXT_ALIGNMENT:
@@ -369,7 +369,7 @@ void Graphics_play (Graphics me, Graphics thee) {
 								0, nrow - 1, y1, y2, minimum, maximum);
 				Melder_free (z);
 			}  break;
-			case SET_COLOUR_SCALE: Graphics_setColourScale (thee, (enum kGraphics_colourScale) get); break;
+			case SET_COLOUR_SCALE: Graphics_setColourScale (thee, (enum kGraphics_colourScale) (int) get); break;
 			case SET_SPECKLE_SIZE: Graphics_setSpeckleSize (thee, get); break;
 			case SPECKLE:
 			{  double x = get, y = get;
